requirements.txt
fastapi==0.87.0
uvicorn==0.20.0
sqlalchemy==1.4.44
mysql-connector-python==8.0.31
python-dotenv==0.17.1
requests==2.28.2